requirements.txt
asgiref==3.3.2
attrs==20.3.0
Automat==20.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
cssselect==1.1.0
defusedxml==0.7.0rc1
Django==3.2.5
django-environ==0.4.5
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
environ==1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
itemadapter==0.2.0
itemloaders==1.0.4
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
lxml==4.6.1
Markdown==3.3.3
MarkupSafe==1.1.1
oauthlib==3.1.0
parsel==1.6.0
pip-tools==5.4.0
Protego==0.1.16
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
python3-openid==3.2.0
pytz==2020.4
queuelib==1.5.0
requests==2.25.0
requests-oauthlib==1.3.0
Scrapy==2.4.1
service-identity==18.1.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==2.0.1
sqlparse==0.4.1
Twisted==20.3.0
uritemplate==3.0.1
urllib3==1.26.2
w3lib==1.22.0
zope.interface==5.2.0
Pillow==8.1.0
stripe==2.65.0
